Reduces components
