+++ app.json
@@ -13,11 +13,7 @@
 			"supportsTablet": true
 		},
 		"android": {
-			"package": "com.crossmint.expo.solana.demo",
-			"adaptiveIcon": {
-				"foregroundImage": "./assets/images/adaptive-icon.png",
-				"backgroundColor": "#ffffff"
-			}
+			"package": "com.crossmint.expo.solana.demo"
 		},
 		"web": {
 			"bundler": "metro",
@@ -25,14 +21,12 @@
 			"favicon": "./assets/images/favicon.png"
 		},
 		"plugins": [
-			"expo-router",
 			[
 				"expo-splash-screen",
 				{
-					"image": "./assets/images/splash-icon.png",
-					"imageWidth": 200,
+					"image": "./assets/images/splash.png",
 					"resizeMode": "contain",
-					"backgroundColor": "#ffffff"
+					"backgroundColor": "#0f0"
 				}
 			]
 		],
