+++ app/_layout.tsx
@@ -1,47 +1,40 @@
 import React from "react";
 import { Stack } from "expo-router";
 import { StyleSheet, View } from "react-native";
-import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import CrossmintProviders from "./providers";
-
-// Import utilities
 import "@/lib/polyfills";
-
-const queryClient = new QueryClient();
 
 export default function RootLayout() {
 	return (
-		<QueryClientProvider client={queryClient}>
-			<CrossmintProviders>
-				<View style={styles.container}>
-					<Stack
-						screenOptions={{
-							headerShown: false,
+		<CrossmintProviders>
+			<View style={styles.container}>
+				<Stack
+					screenOptions={{
+						headerShown: false,
+						contentStyle: {
+							flex: 1,
+						},
+					}}
+				>
+					<Stack.Screen
+						name="login"
+						options={{
 							contentStyle: {
-								flex: 1,
+								backgroundColor: "#FFFFFF",
 							},
 						}}
-					>
-						<Stack.Screen
-							name="login"
-							options={{
-								contentStyle: {
-									backgroundColor: "#FFFFFF",
-								},
-							}}
-						/>
-						<Stack.Screen
-							name="wallet"
-							options={{
-								contentStyle: {
-									backgroundColor: "#F8FAFC",
-								},
-							}}
-						/>
-					</Stack>
-				</View>
-			</CrossmintProviders>
-		</QueryClientProvider>
+					/>
+					<Stack.Screen
+						name="wallet"
+						options={{
+							contentStyle: {
+								backgroundColor: "#F8FAFC",
+							},
+						}}
+					/>
+				</Stack>
+			</View>
+		</CrossmintProviders>
 	);
 }
