--- components/icons/CrossmintLeaf.tsx
@@ -1,31 +1,0 @@
-import * as React from "react";
-import type { ViewStyle } from "react-native";
-import { Svg, Path } from "react-native-svg";
-
-type Props = {
-	style?: ViewStyle;
-};
-
-export default function CrossmintLeaf({ style }: Props) {
-	return (
-		<Svg
-			width={361}
-			height={52}
-			viewBox="0 0 361 52"
-			accessibilityLabel="Crossmint leaf"
-			style={style}
-		>
-			<Path
-				fill-rule="evenodd"
-				clip-rule="evenodd"
-				d="M118.94 26.933c-1.258-.644-2.937-.86-3.922-.934 1.341-.1 3.967-.464 5.067-1.81 1.72-1.289 1.638-5.646 1.638-5.646s-4.128-.455-5.931 1.6c-1.12 1.115-1.467 3.052-1.568 4.401-.102-1.347-.449-3.286-1.569-4.4-1.802-2.058-5.931-1.6-5.931-1.6s-.05 2.716.775 4.494c.407.876 1.15 1.59 2.008 2.03 1.258.643 2.937.86 3.923.933-.986.073-2.665.29-3.923.934-.858.44-1.601 1.153-2.008 2.029-.825 1.776-.775 4.493-.775 4.493s4.129.456 5.931-1.6c1.12-1.115 1.467-3.054 1.569-4.4.101 1.346.448 3.285 1.568 4.4 1.803 2.056 5.931 1.6 5.931 1.6s.052-2.717-.775-4.495c-.406-.876-1.15-1.59-2.008-2.03Zm.138 3.81c-.021-.005-2.176-.619-4.942-4.313a34.005 34.005 0 0 0-4.789 4.596l-.087.103.034-.13c.006-.023.637-2.243 4.485-5.07-.372-.55-1.246-1.82-4.621-4.808l-.075-.067.1.014c.071.01 1.806.283 4.974 4.366 0 0 .04.056.114.153.446-.308 1.739-1.226 4.646-4.51l.067-.075-.014.1c-.011.07-.283 1.797-4.34 4.953a33.65 33.65 0 0 0 4.476 4.635l.104.089-.13-.035h-.002Z"
-				fill="#A4AFB2"
-			/>
-
-			<Path
-				d="M127.78 31v-9.454h3.545c.727 0 1.336.135 1.829.406.495.27.869.643 1.121 1.117.256.47.383 1.006.383 1.606 0 .607-.127 1.145-.383 1.616-.255.471-.632.842-1.131 1.113-.498.268-1.112.401-1.842.401h-2.349v-1.408h2.118c.425 0 .773-.073 1.044-.221.271-.148.471-.351.6-.61.132-.258.198-.555.198-.89 0-.336-.066-.631-.198-.887a1.347 1.347 0 0 0-.605-.596c-.271-.144-.62-.216-1.048-.216h-1.569V31h-1.713Zm10.504.139c-.693 0-1.293-.153-1.801-.457a3.107 3.107 0 0 1-1.181-1.28c-.277-.547-.416-1.187-.416-1.92 0-.732.139-1.374.416-1.925a3.1 3.1 0 0 1 1.181-1.283c.508-.305 1.108-.457 1.801-.457.692 0 1.292.152 1.8.457.508.304.9.732 1.177 1.283.28.551.42 1.193.42 1.925 0 .733-.14 1.373-.42 1.92a3.072 3.072 0 0 1-1.177 1.28c-.508.304-1.108.457-1.8.457Zm.009-1.34c.375 0 .689-.102.942-.309.252-.209.44-.489.563-.84a3.45 3.45 0 0 0 .189-1.172c0-.434-.063-.827-.189-1.178a1.865 1.865 0 0 0-.563-.844c-.253-.21-.567-.314-.942-.314-.385 0-.705.104-.96.314-.253.209-.442.49-.568.844-.123.351-.185.744-.185 1.178 0 .43.062.821.185 1.172.126.351.315.631.568.84.255.207.575.31.96.31ZM143.58 31l-2.003-7.09h1.703l1.247 4.985h.064l1.274-4.986h1.685l1.275 4.958h.069l1.228-4.958h1.708L149.822 31h-1.741l-1.329-4.792h-.097L145.325 31h-1.745Zm11.573.139c-.711 0-1.325-.148-1.842-.444a3.003 3.003 0 0 1-1.187-1.265c-.277-.547-.415-1.192-.415-1.934 0-.73.138-1.37.415-1.92a3.14 3.14 0 0 1 1.173-1.293c.502-.31 1.091-.466 1.768-.466.437 0 .849.07 1.237.212a2.79 2.79 0 0 1 1.034.646c.302.293.539.665.711 1.118.173.449.259.985.259 1.606v.513h-5.812v-1.127h4.21a1.79 1.79 0 0 0-.208-.854 1.508 1.508 0 0 0-.568-.595 1.595 1.595 0 0 0-.84-.217c-.342 0-.642.083-.9.249-.259.163-.46.378-.605.646a1.866 1.866 0 0 0-.217.873v.983c0 .412.076.766.226 1.062.151.292.362.517.633.674.271.154.588.23.951.23.243 0 .463-.033.66-.1.197-.072.368-.175.512-.31.145-.136.254-.303.328-.503l1.561.175a2.44 2.44 0 0 1-.564 1.08 2.743 2.743 0 0 1-1.052.711c-.428.166-.917.25-1.468.25ZM158.94 31v-7.09h1.62v1.18h.074c.129-.408.351-.724.665-.945a1.828 1.828 0 0 1 1.084-.337c.093 0 .196.004.31.013.117.006.214.017.291.033v1.537a1.779 1.779 0 0 0-.337-.065 3.11 3.11 0 0 0-.439-.032 1.69 1.69 0 0 0-.822.198 1.457 1.457 0 0 0-.775 1.34V31h-1.671Zm7.221.139c-.711 0-1.325-.148-1.842-.444a3 3 0 0 1-1.186-1.265c-.277-.547-.416-1.192-.416-1.934 0-.73.139-1.37.416-1.92.28-.554.671-.985 1.172-1.293.502-.31 1.091-.466 1.768-.466.437 0 .85.07 1.238.212.39.139.735.354 1.034.646.301.293.538.665.711 1.118.172.449.258.985.258 1.606v.513h-5.812v-1.127h4.21a1.79 1.79 0 0 0-.208-.854 1.506 1.506 0 0 0-.567-.595 1.596 1.596 0 0 0-.841-.217 1.63 1.63 0 0 0-.9.249c-.258.163-.46.378-.605.646a1.865 1.865 0 0 0-.216.873v.983c0 .412.075.766.226 1.062.151.292.361.517.632.674.271.154.588.23.951.23.243 0 .463-.033.66-.1a1.47 1.47 0 0 0 .513-.31c.144-.136.254-.303.328-.503l1.56.175a2.438 2.438 0 0 1-.563 1.08 2.747 2.747 0 0 1-1.053.711c-.428.166-.917.25-1.468.25Zm6.418-.014a2.68 2.68 0 0 1-1.495-.43c-.441-.286-.788-.701-1.044-1.246-.255-.545-.383-1.207-.383-1.985 0-.788.129-1.453.388-1.994.261-.545.614-.956 1.057-1.233.443-.28.937-.42 1.482-.42.415 0 .757.07 1.025.212.268.139.48.306.637.503.157.194.278.377.365.55h.069v-3.537h1.676V31h-1.644v-1.117h-.101a2.899 2.899 0 0 1-.374.55c-.163.19-.379.353-.647.489-.267.135-.604.203-1.011.203Zm.467-1.371c.354 0 .655-.096.904-.287.25-.194.439-.463.568-.808.13-.344.194-.746.194-1.204 0-.459-.064-.858-.194-1.196a1.702 1.702 0 0 0-.563-.79c-.246-.187-.549-.281-.909-.281-.373 0-.684.097-.933.29a1.761 1.761 0 0 0-.563.804 3.375 3.375 0 0 0-.189 1.172c0 .444.063.84.189 1.187.129.345.318.617.568.817.252.197.561.296.928.296ZM179.875 31v-9.454h1.671v3.536h.069c.086-.173.208-.356.365-.55.157-.197.369-.364.637-.503.268-.142.609-.212 1.025-.212.548 0 1.042.14 1.482.42.443.277.794.688 1.052 1.233.262.541.393 1.206.393 1.994 0 .778-.128 1.44-.383 1.985-.256.545-.604.96-1.044 1.246a2.68 2.68 0 0 1-1.495.43c-.407 0-.744-.068-1.011-.203a2.056 2.056 0 0 1-.647-.49 3.212 3.212 0 0 1-.374-.55h-.097V31h-1.643Zm1.639-3.546c0 .46.064.86.194 1.205.132.345.321.614.567.808.25.191.551.287.905.287.37 0 .679-.099.928-.296.249-.2.437-.472.563-.817.13-.348.194-.743.194-1.186 0-.44-.063-.831-.189-1.173a1.761 1.761 0 0 0-.563-.803c-.249-.194-.56-.291-.933-.291-.357 0-.66.094-.909.282-.249.187-.439.45-.568.789-.126.338-.189.737-.189 1.195Zm6.493 6.205a3.372 3.372 0 0 1-1.089-.175l.387-1.302c.243.07.46.104.651.101a.796.796 0 0 0 .503-.18c.148-.114.273-.304.374-.572l.143-.383-2.571-7.239h1.773l1.634 5.355h.074l1.639-5.355h1.777l-2.839 7.95a3.188 3.188 0 0 1-.526.965c-.219.27-.487.477-.804.618-.314.145-.689.217-1.126.217ZM199.479 25.467c-1.067 0-2.013.796-2.013 2.268 0 1.472.946 2.298 2.043 2.298 1.141 0 1.667-.796 1.832-1.352l1.533.556c-.346 1.142-1.442 2.389-3.365 2.389-2.148 0-3.785-1.668-3.785-3.89 0-2.254 1.637-3.876 3.74-3.876 1.968 0 3.049 1.232 3.349 2.403l-1.562.57c-.165-.645-.661-1.366-1.772-1.366ZM208.113 24.04v1.773a3.8 3.8 0 0 0-.571-.046c-1.351 0-1.967.782-1.967 2.148V31.4h-1.743v-7.315h1.698v1.172c.345-.796 1.156-1.262 2.118-1.262.21 0 .39.03.465.045ZM212.11 30.063c1.082 0 2.043-.796 2.043-2.328 0-1.517-.961-2.298-2.043-2.298-1.066 0-2.042.781-2.042 2.298s.976 2.328 2.042 2.328Zm0-6.203c2.194 0 3.801 1.637 3.801 3.875 0 2.253-1.607 3.89-3.801 3.89-2.178 0-3.785-1.637-3.785-3.89 0-2.238 1.607-3.875 3.785-3.875ZM216.495 29.418l1.517-.421c.06.66.556 1.247 1.487 1.247.721 0 1.097-.39 1.097-.842 0-.39-.27-.69-.856-.81l-1.082-.241c-1.292-.285-1.967-1.127-1.967-2.133 0-1.277 1.171-2.358 2.688-2.358 2.043 0 2.704 1.322 2.824 2.043l-1.472.42c-.06-.42-.375-1.096-1.352-1.096-.616 0-1.051.39-1.051.84 0 .391.285.677.736.767l1.111.225c1.382.3 2.103 1.142 2.103 2.223 0 1.112-.901 2.343-2.764 2.343-2.117 0-2.929-1.381-3.019-2.207ZM222.877 29.418l1.517-.421c.06.66.556 1.247 1.487 1.247.721 0 1.096-.39 1.096-.842 0-.39-.27-.69-.856-.81l-1.081-.241c-1.292-.285-1.968-1.127-1.968-2.133 0-1.277 1.172-2.358 2.689-2.358 2.043 0 2.703 1.322 2.824 2.043l-1.472.42c-.06-.42-.376-1.096-1.352-1.096-.616 0-1.052.39-1.052.84 0 .391.286.677.736.767l1.112.225c1.382.3 2.103 1.142 2.103 2.223 0 1.112-.902 2.343-2.764 2.343-2.118 0-2.929-1.381-3.019-2.207ZM231.496 31.4h-1.727v-7.315h1.667v.931c.421-.75 1.337-1.141 2.163-1.141.962 0 1.803.435 2.193 1.292.571-.962 1.427-1.292 2.374-1.292 1.321 0 2.583.841 2.583 2.779V31.4h-1.682V26.94c0-.856-.436-1.502-1.412-1.502-.916 0-1.517.721-1.517 1.622V31.4h-1.713V26.94c0-.841-.42-1.502-1.412-1.502-.931 0-1.517.69-1.517 1.622V31.4ZM243.965 31.4h-1.727v-7.315h1.727V31.4Zm-1.997-9.883c0-.631.51-1.142 1.126-1.142.631 0 1.142.51 1.142 1.142 0 .63-.511 1.126-1.142 1.126a1.124 1.124 0 0 1-1.126-1.126ZM247.267 27.18v4.22h-1.743v-7.315h1.698v.976c.48-.84 1.352-1.186 2.163-1.186 1.787 0 2.643 1.292 2.643 2.899V31.4h-1.742v-4.326c0-.901-.406-1.622-1.502-1.622-.992 0-1.517.766-1.517 1.727ZM255.744 21.847v2.238h1.517v1.547h-1.517v3.395c0 .646.286.916.931.916.241 0 .526-.045.601-.06v1.442c-.105.045-.435.165-1.066.165-1.352 0-2.193-.81-2.193-2.178v-3.68h-1.352v-1.547h.375c.782 0 1.127-.496 1.127-1.142v-1.096h1.577Z"
-				fill="#A4AFB2"
-			/>
-		</Svg>
-	);
-}
