--- components/icons/CrossmintLogo.tsx
@@ -1,71 +1,0 @@
-import * as React from "react";
-import type { ViewStyle } from "react-native";
-import {
-	Svg,
-	Defs,
-	LinearGradient,
-	Stop,
-	G,
-	Path,
-	ClipPath,
-} from "react-native-svg";
-
-type Props = {
-	style?: ViewStyle;
-};
-
-export default function CrossmintLogo({ style }: Props) {
-	return (
-		<Svg
-			width={127}
-			height={24}
-			viewBox="0 0 552 105"
-			fill="none"
-			accessibilityLabel="Crossmint logo"
-			style={style}
-		>
-			<G clipPath="url(#a)">
-				<Path
-					d="M447.779 12.33c0-4.08 3.3-7.38 7.284-7.38s7.38 3.3 7.38 7.38-3.3 7.284-7.38 7.284-7.284-3.204-7.284-7.284ZM541.175 14.586v15.24h10.332v10.536h-10.332v23.112c0 4.404 1.944 6.24 6.348 6.24 1.632 0 3.576-.312 4.092-.408v9.816c-.72.312-2.964 1.128-7.26 1.128-9.204 0-14.928-5.52-14.928-14.832V40.362h-9.204V29.826h10.224v-15.24h10.74-.012ZM135.635 67.89c2.28 3.948 5.388 7.092 9.336 9.432 4.008 2.28 8.508 3.42 13.5 3.42 3.636 0 6.996-.612 10.08-1.848 3.084-1.296 5.736-3.048 7.956-5.268 2.22-2.28 3.816-4.836 4.812-7.668l-10.632-4.812c-.924 2.592-2.46 4.656-4.62 6.192-2.16 1.536-4.68 2.316-7.584 2.316-2.652 0-5.028-.648-7.116-1.944-2.028-1.296-3.636-3.084-4.812-5.364-1.176-2.28-1.752-4.896-1.752-7.86 0-2.964.588-5.58 1.752-7.86 1.176-2.28 2.772-4.068 4.812-5.364 2.1-1.296 4.464-1.944 7.116-1.944 2.832 0 5.328.768 7.488 2.316 2.22 1.536 3.792 3.576 4.716 6.096l10.632-4.62c-.984-3.024-2.616-5.604-4.896-7.764-2.22-2.22-4.872-3.948-7.956-5.172-3.084-1.296-6.408-1.944-9.984-1.944-4.992 0-9.492 1.14-13.5 3.42-3.948 2.28-7.056 5.388-9.336 9.336-2.28 3.948-3.42 8.412-3.42 13.404 0 4.992 1.14 9.492 3.42 13.5h-.012ZM199.247 29.346h-11.376v50.292h12.108V51.63c0-3.816 1.044-6.78 3.144-8.88 2.1-2.16 4.872-3.24 8.316-3.24h4.344V28.782h-2.964c-3.516 0-6.528.744-9.06 2.22-1.896 1.14-3.396 2.94-4.524 5.412v-7.08l.012.012Z"
-					fill="#222"
-				/>
-				<Path
-					fillRule="evenodd"
-					clipRule="evenodd"
-					d="M244.079 28.29c14.928 0 25.884 11.148 25.884 26.388 0 15.24-10.944 26.496-25.884 26.496s-25.776-11.148-25.776-26.496c0-15.348 10.944-26.388 25.776-26.388Zm0 42.24c7.368 0 13.908-5.424 13.908-15.852s-6.552-15.648-13.908-15.648-13.908 5.316-13.908 15.648 6.648 15.852 13.908 15.852Z"
-					fill="#222"
-				/>
-				<Path
-					d="m284.267 63.27-10.332 2.868c.612 5.628 6.132 15.036 20.556 15.036 12.684 0 18.816-8.388 18.816-15.96 0-7.572-4.908-13.092-14.316-15.144l-7.572-1.536c-3.072-.612-5.016-2.556-5.016-5.22 0-3.072 2.964-5.724 7.164-5.724 6.648 0 8.796 4.608 9.204 7.464l10.02-2.868c-.816-4.908-5.316-13.908-19.224-13.908-10.332 0-18.312 7.368-18.312 16.056 0 6.852 4.608 12.576 13.404 14.52l7.368 1.632c3.984.816 5.832 2.868 5.832 5.52 0 3.072-2.556 5.724-7.464 5.724-6.336 0-9.72-3.984-10.128-8.484v.024ZM317.291 66.138l10.332-2.868c.408 4.5 3.78 8.484 10.128 8.484 4.908 0 7.464-2.664 7.464-5.724 0-2.664-1.836-4.704-5.832-5.52l-7.368-1.632c-8.796-1.944-13.404-7.668-13.404-14.52 0-8.7 7.98-16.056 18.312-16.056 13.908 0 18.408 9 19.224 13.908l-10.02 2.868c-.408-2.868-2.556-7.464-9.204-7.464-4.188 0-7.164 2.664-7.164 5.724 0 2.664 1.944 4.608 5.016 5.22l7.572 1.536c9.408 2.04 14.316 7.776 14.316 15.144s-6.132 15.96-18.816 15.96c-14.424 0-19.944-9.408-20.556-15.036v-.024ZM375.755 29.346h-11.376v50.292h12.108V50.142c0-2.22.396-4.128 1.2-5.736.804-1.608 1.944-2.832 3.42-3.696 1.476-.924 3.18-1.392 5.088-1.392s3.672.468 5.088 1.392c1.476.864 2.616 2.1 3.42 3.696.804 1.608 1.2 3.516 1.2 5.736v29.496h12.108V50.142c0-2.22.396-4.128 1.2-5.736.804-1.608 1.944-2.832 3.42-3.696 1.476-.924 3.18-1.392 5.088-1.392 2.028 0 3.756.468 5.172 1.392 1.416.864 2.532 2.1 3.324 3.696.804 1.608 1.2 3.516 1.2 5.736v29.496h12.108V47.286c0-3.756-.804-7.056-2.4-9.888-1.596-2.892-3.816-5.148-6.66-6.744-2.772-1.608-5.952-2.4-9.528-2.4-4.008 0-7.548 1.02-10.632 3.048-1.98 1.272-3.648 2.964-4.992 5.076-1.104-1.896-2.64-3.528-4.62-4.884-3.144-2.16-6.78-3.24-10.908-3.24-3.756 0-7.02.888-9.804 2.676-1.812 1.164-3.228 2.736-4.248 4.692v-6.264l.024-.012ZM461.003 79.638h-11.76V29.826h11.76v49.812ZM471.647 29.814v49.812h12V50.418c0-2.196.432-4.092 1.284-5.676.852-1.584 2.04-2.808 3.576-3.66 1.524-.912 3.264-1.368 5.22-1.368s3.756.456 5.22 1.368c1.524.852 2.712 2.076 3.576 3.66.852 1.584 1.284 3.48 1.284 5.676v29.208h12V47.574c0-3.72-.792-6.996-2.376-9.804a16.743 16.743 0 0 0-6.684-6.684c-2.808-1.584-6.048-2.376-9.708-2.376s-6.66.792-9.336 2.376a12.716 12.716 0 0 0-4.764 4.836v-6.108h-11.292Z"
-					fill="#222"
-				/>
-				<Path
-					fillRule="evenodd"
-					clipRule="evenodd"
-					d="M84.61 58.914c-8.651-4.428-20.195-5.916-26.975-6.42 9.228-.684 27.288-3.192 34.848-12.456 11.832-8.856 11.268-38.82 11.268-38.82S75.359-1.914 62.963 12.222C55.259 19.89 52.87 33.21 52.175 42.486c-.696-9.264-3.084-22.596-10.788-30.264C28.99-1.926.599 1.218.599 1.218S.25 19.902 5.927 32.13c2.796 6.024 7.908 10.932 13.812 13.956 8.652 4.428 20.196 5.916 26.976 6.42-6.78.504-18.324 1.992-26.976 6.42-5.904 3.024-11.016 7.932-13.812 13.956-5.676 12.216-5.328 30.9-5.328 30.9s28.392 3.132 40.788-11.004c7.704-7.668 10.092-21 10.788-30.264.696 9.264 3.084 22.596 10.788 30.264 12.396 14.136 40.788 11.004 40.788 11.004s.36-18.684-5.328-30.912c-2.796-6.024-7.908-10.932-13.812-13.956Zm.949 26.208c-.144-.036-14.964-4.26-33.984-29.664-5.58 4.356-19.596 15.876-32.94 31.608l-.6.708.24-.9c.036-.156 4.38-15.42 30.84-34.86-2.556-3.792-8.568-12.516-31.776-33.072l-.516-.456.684.096c.492.072 12.42 1.944 34.212 30.024 0 0 .276.384.78 1.056 3.072-2.124 11.964-8.436 31.956-31.02l.456-.516-.096.684c-.072.492-1.944 12.36-29.844 34.068 5.268 6.648 16.164 19.488 30.78 31.872l.72.612-.9-.24h-.012Z"
-					fill="url(#b)"
-				/>
-			</G>
-			<Defs>
-				<LinearGradient
-					id="b"
-					x1=".743"
-					y1="1.074"
-					x2="103.607"
-					y2="103.926"
-					gradientUnits="userSpaceOnUse"
-				>
-					<Stop stopColor="#5EDD4D" />
-					<Stop offset="1" stopColor="#05CE6C" />
-				</LinearGradient>
-				<ClipPath id="a">
-					<Path
-						fill="#fff"
-						transform="translate(.6 .918)"
-						d="M0 0h551.004v103.164H0z"
-					/>
-				</ClipPath>
-			</Defs>
-		</Svg>
-	);
-}
