+++ package.json
@@ -21,7 +21,6 @@
 		"@react-navigation/native": "^7.0.19",
 		"@solana/spl-token": "^0.4.13",
 		"@solana/web3.js": "^1.98.0",
-		"@tanstack/react-query": "^5.71.0",
 		"buffer": "^6.0.3",
 		"expo": "~52.0.42",
 		"expo-blur": "~14.0.3",
@@ -52,7 +51,6 @@
 	},
 	"devDependencies": {
 		"@babel/core": "^7.26.10",
-		"@tanstack/eslint-plugin-query": "^5.68.0",
 		"@types/jest": "^29.5.12",
 		"@types/react": "~18.3.20",
 		"@types/react-test-renderer": "^18.3.1",
